+++ src/tkc/fscript.c
@@ -2275,6 +2275,14 @@ static ret_t func_unset(fscript_t* fscript, fscript_args_t* args, value_t* resul
   return RET_OK;
 }
 
+static ret_t func_sleep_ms(fscript_t* fscript, fscript_args_t* args, value_t* result) {
+  FSCRIPT_FUNC_CHECK(args->size == 1, RET_BAD_PARAMS);
+  sleep_ms(value_uint32(args->args));
+  value_set_bool(result, TRUE);
+
+  return RET_OK;
+}
+
 typedef struct _func_entry_t {
   const char* name;
   fscript_func_t func;
@@ -2365,6 +2373,7 @@ static const func_entry_t s_builtin_funcs[] = {
     {"*", func_mul, 2},
     {"-", func_sub, 2},
     {"+", func_sum, 8},
+    {"sleep_ms", func_sleep_ms, 1}
 };
 
 static general_factory_t* s_global_funcs;